mediana arreglada

+++ libft/src/ft_to/ft_itoa.c
@@ -29,7 +29,7 @@ static int	ft_len(int n)
 	return (i);
 }
 
-char	*ft_itoa(int n)
+char	*ft_itoa(long n)
 {
 	char	*str;
 	int		len;
